requirements.txt
pillow==9.2.0
pygame==2.0.1
pyyaml==6.0
stable_baselines3==1.1.0a7
ray[rllib]==1.10.0
gym==0.21.0
jupyter==1.0.0
jupyterlab==4.2.1
